requirements.txt
flask-cors
python-dotenv
openai==0.28.0
flask-jwt-extended
flask-mail 